requirements.txt
tqdm
wget
scipy
numpy
opencv-python
h5py
scikit-video
tensorflow-gpu
matplotlib
zmq